Don't create cache when PHPSESSID cookie exists
Need this as while clients with the PHPSESSID cookie are never served a
cache, they still create new caches meaning that clients that does not
have PHPSESSID cookie set may be served a cached page from a logged in
user, spelling all kinds of trouble (ie. users appearing logged in as an
administrator).